Add files via upload
The Weather App is a user-friendly web application designed to provide real-time weather updates based on user input. It utilizes APIs to fetch current weather data and presents it in an organized and visually appealing format. This project was developed using HTML, CSS, and JavaScript, with integration of the OpenWeatherMap API to retrieve live weather data.